+++ docs/news.rst
@@ -11,6 +11,11 @@ Added
 
 - Add a :ref:`jobs_dir` setting, to create a `job directory <https://doc.scrapy.org/en/latest/topics/jobs.html>`__ for each crawl, to persist requests to disk.
 
+Changed
+~~~~~~~
+
+- The :ref:`cancel.json` webservice accepts an integer for the ``signal`` parameter.
+
 Fixed
 ~~~~~